AbstractIOHandler: Vtable ODR
Make `AbstractIOHandler` a purely virtual base class
with only inline member functions
and move out `DummyIOHandler` into separate header.

That avoids spilling a vtable for it in every translation unit
that includes the header.
