+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Bug Fixes
 
 - spurious MPI C++11 API usage in ParallelIOTest removed #396
 - use-after-free in SerialIOTest #409
+- fix ODR issue in ADIOS1 backend corrupting the ``AbstractIOHandler`` vtable #415
 
 Other
 """""
