+++ include/openPMD/IO/AbstractIOHandlerImpl.hpp
@@ -20,6 +20,7 @@
  */
 #pragma once
 
+#include "openPMD/IO/AbstractIOHandler.hpp"
 #include "openPMD/IO/IOTask.hpp"
 
 #include <future>
@@ -27,16 +28,92 @@
 
 namespace openPMD
 {
-class AbstractIOHandler;
+// class AbstractIOHandler;
 class Writable;
 
 class AbstractIOHandlerImpl
 {
 public:
-  AbstractIOHandlerImpl(AbstractIOHandler*);
-  virtual ~AbstractIOHandlerImpl();
+    AbstractIOHandlerImpl(AbstractIOHandler *handler)
+    : m_handler{handler}
+    { }
 
-  virtual std::future< void > flush();
+    virtual ~AbstractIOHandlerImpl() = default;
+
+    virtual std::future< void > flush()
+    {
+        while( !(*m_handler).m_work.empty() )
+        {
+            IOTask& i = (*m_handler).m_work.front();
+            try
+            {
+                switch( i.operation )
+                {
+                    using O = Operation;
+                    case O::CREATE_FILE:
+                        createFile(i.writable, *dynamic_cast< Parameter< Operation::CREATE_FILE >* >(i.parameter.get()));
+                        break;
+                    case O::CREATE_PATH:
+                        createPath(i.writable, *dynamic_cast< Parameter< O::CREATE_PATH >* >(i.parameter.get()));
+                        break;
+                    case O::CREATE_DATASET:
+                        createDataset(i.writable, *dynamic_cast< Parameter< O::CREATE_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::EXTEND_DATASET:
+                        extendDataset(i.writable, *dynamic_cast< Parameter< O::EXTEND_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::OPEN_FILE:
+                        openFile(i.writable, *dynamic_cast< Parameter< O::OPEN_FILE >* >(i.parameter.get()));
+                        break;
+                    case O::OPEN_PATH:
+                        openPath(i.writable, *dynamic_cast< Parameter< O::OPEN_PATH >* >(i.parameter.get()));
+                        break;
+                    case O::OPEN_DATASET:
+                        openDataset(i.writable, *dynamic_cast< Parameter< O::OPEN_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::DELETE_FILE:
+                        deleteFile(i.writable, *dynamic_cast< Parameter< O::DELETE_FILE >* >(i.parameter.get()));
+                        break;
+                    case O::DELETE_PATH:
+                        deletePath(i.writable, *dynamic_cast< Parameter< O::DELETE_PATH >* >(i.parameter.get()));
+                        break;
+                    case O::DELETE_DATASET:
+                        deleteDataset(i.writable, *dynamic_cast< Parameter< O::DELETE_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::DELETE_ATT:
+                        deleteAttribute(i.writable, *dynamic_cast< Parameter< O::DELETE_ATT >* >(i.parameter.get()));
+                        break;
+                    case O::WRITE_DATASET:
+                        writeDataset(i.writable, *dynamic_cast< Parameter< O::WRITE_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::WRITE_ATT:
+                        writeAttribute(i.writable, *dynamic_cast< Parameter< O::WRITE_ATT >* >(i.parameter.get()));
+                        break;
+                    case O::READ_DATASET:
+                        readDataset(i.writable, *dynamic_cast< Parameter< O::READ_DATASET >* >(i.parameter.get()));
+                        break;
+                    case O::READ_ATT:
+                        readAttribute(i.writable, *dynamic_cast< Parameter< O::READ_ATT >* >(i.parameter.get()));
+                        break;
+                    case O::LIST_PATHS:
+                        listPaths(i.writable, *dynamic_cast< Parameter< O::LIST_PATHS >* >(i.parameter.get()));
+                        break;
+                    case O::LIST_DATASETS:
+                        listDatasets(i.writable, *dynamic_cast< Parameter< O::LIST_DATASETS >* >(i.parameter.get()));
+                        break;
+                    case O::LIST_ATTS:
+                        listAttributes(i.writable, *dynamic_cast< Parameter< O::LIST_ATTS >* >(i.parameter.get()));
+                        break;
+                }
+            } catch (unsupported_data_error&)
+            {
+                (*m_handler).m_work.pop();
+                throw;
+            }
+            (*m_handler).m_work.pop();
+        }
+        return std::future< void >();
+    }
 
   /** Create a new file in physical storage, possibly overriding an existing file.
    *
